[6.2][Commands] Migrate: Avoid duplicate fix-its and manifest updates when… (#8929)
… building target for both host and destination

- Explanation:

This is a limited cherry-pick of
https://github.com/swiftlang/swift-package-manager/pull/8888

For example if a target is a dependency of a plugin we build it for host
and destination and produce the same fix-it twice and attempt to update
manifest twice because uniquing is done on the ModuleBuildDescriptions.

  The changes do the following:

- Coalesce all the diagnostic files for the same target regardless of
host or destination build to make it possible for`SwiftFixIt` to filter
duplicate fix-its;
- Unique based on module identity to prevent duplicate setting
insertion.

- Resolves: rdar://155569285

- Main Branch PR:
https://github.com/swiftlang/swift-package-manager/pull/8888

- Risk: Very Low. This is a narrow fix for `swift package migrate`
command.
 
- Reviewed By: Me and @AnthonyLatsis reviewed the work in the main PR.

- Testing: Added new test-cases to the suite.