Tests: Update isAmazonLinux2() logic
The ProcessInfo extension to determine if a host is running on Amazon
Linux 2 was slightly flawed as it did not take into account the
double-quote (") surrounding the values.

Update the logic to look at "PRETTY_NAME" since the ID can be used in
multiple Amazon Linux OS'es, and use exact logic match.

issue: rdar://155612963