fix deprecation to introduction for publicHeadersDirectoryURL property (#8870)
Deprecation looks "incorrect" as it was added and deprecated at the same
as the property `publicHeadersDirectory`, which I believe this was meant
to replace.

### Motivation:

While updating the documentation, this stood out as added and
deprecated, which seemed like a mistake.

### Modifications:

switched the deprecation to an introduction for the publicHeadersDirectoryURL property

### Result:

property is undeprecated