Tests: Do not use mix of XCTest and Swift Testing
The fixture(..) helper function is being used in XCTest and Swift
Testing tests. However, there was a code path that could lead to calling
XCTFail(), which is a no-op in Swift Testing.

This change rename the functio to fixtureXCTest, and marks it as
deprecated with a note indicating to migrate the test to Swift Testing.
In addition, create a Swift Testing equivalent of the fixture class that
calls the relevant APIs.