[BuildPlan] Fix `traverseModules` to account for tests with direct ma… (#8940)
…cro dependencies

### Motivation:

Currently there is a difference in behavior between
`BuildPlan.computeDestinations` and `BuildPlan.traverseModules`
where the latter always returns test targets as built for the "target".
This is incorrect if i.e. targets have direct macro
dependencies.

### Modifications:

This change aligns `traverseModules` with `computeDestinations` in
regard to test targets that are built for the host.
And let's test targets be discovered through an aggregate produce that
also informs their destination.

### Result:

The fix helps sourcekit-lsp to correctly identify test targets.

Resolves: https://github.com/swiftlang/sourcekit-lsp/issues/2205