Docs/package plugin api docs (#8901)
Creates a documentation catalog and edits the content to provide public
documentation for the Package Plugin API

resolves #8827

### Motivation:

The Package Plugin API is required when creating swift package manager
plugins, but doesn't have publicly hosted documentation today. This
resolves that by creating the documentation catalog, adding curation
(organization), and editing the documentation content to match.

### Modifications:

- creates a documentation catalog
- provides curation (extension) DocC pages to provide organization to
the module and to each of the public types in the module
- edits to the documentation comments to align stylistically with DocC
conventions.

### Result:

The potential to build a documentation archive for the Package Plugin
API.

Preview the contents of this PR using the command:
```
swift package --disable-sandbox preview-documentation --target PackagePlugin
```

---------

Co-authored-by: Bri Peticca <bripeticca@gmail.com>