Run Command: Add SwiftBuild repl support (#8971)
The `swift run --repl` was explicitly using the Native build system.
This change modified the `swift run --repl` command to respect the
`--build-system <arg>` command line option. The change introduces a new
build output request `replArguments`. When provided, the build system is
responsible for providing the arguments required for the REPL in the
BuildResult.
    
The Swift Run Command will inspect the build result for this property.
If it's unavailable, the command provides an error indicating repl
support is unavailable.
    
A caveat, more work is required to get proper REPL integration with the
Package. At the moment, the System Library paths are not provided by the
SwiftBuild System when a repl session is requested via the `run`
command.
    
Relates to: #8846
issue: rdar://153822861

Depends on : #8942