New technique for resolving module name conflicts to module alias docs
Conflicts can arise due to duplicate module names within a package
graph. Module aliases themselves if they are not globally unique can
introduce further conflicts. The documentation for the module aliases
feature can describe a technique for the naming that can help to
reduce the likelihood of further collisions by recommending
the use of the package identifier as a qualifier to the alias. As other
packages in the graph encounter the collision, this system can help
for there to be a broader agreement on the new name, which should
lessen the cases where module alias chaining is needed.

Update the correct ModuleAliases.md file so that the changes will be
shown in the new documentation system.