Only fetch swiftc target info when needed.
We added fetching of the swift compiler version using
swiftc -print-target-info in the init of UserToolchain so
that it can be passed to the PrebuiltsManager to calculate
the manifest and prebuilts zip file names.

However, there are cases where we pass in the triple for the
current SDK which was intended to avoid the call to swiftc.
We now call it all the time which causes an extra call to swiftc
that isn't always needed.

To address this, we make the calculation of the target info lazy
and move the call to fetch the swift version to the point we need
it in the PrebuiltsManager. If prebuilts aren't required,
the call is avoided.