Add build prebuilts support for any version of swift-syntax.
Removes the hardcoding of the contents of the prebuilt libraries.
Adds a --version argument to specify which versions of swift-syntax
you want to build for. You can have multiple of these. Also removes
the hardcoding of the list of products that go into the library.
Instead we ask the package for it's list of library products and
add them all.