Fix for .when(traits:) condition not working for multiple traits
The culprit here is the fact that we were essentially doing an
AND boolean check across all the traits in the target dependency
condition, when we should be doing an OR boolean check.

If at least one trait is enabled in the list of traits in
the target dependency condition, then the target dependecy
should be considered enabled as well.