Add Snippet support with SwiftBuild
Snippets are treated as executable targets with the native build system.

This change updates the PIF Builder to support snippet, giving Snippet
support with the Swift Build build system.

Also update the `testCommandPluginCancellation` in PluginTests as it was
not properly handling the task cancellations.

Depends on: https://github.com/swiftlang/swift-build/pull/775
Fixes: #9040
issue: rdar://158630024
issue: rdar://147705448