Add bootstrap option to install without building.
This allows the build action to be decoupled from the install action, so
packaging systems that expect to be able to do discrete installation
steps post-build can do so cheaply, without having to effectively start
the build over from scratch.

This is only intended solely for that use-case, so this is not wired up
to other part of the swiftpm build process and defaults to being
disabled.