[Commands] Update `AddTarget` command to add auxiliary files for new … (#9043)
…targets

### Motivation:

The command currently relies on the refactoring action to supply it with
auxiliary files to add, but we'd like refactoring actions to only be
responsible for manifest updates and everything else should be done
either by a command itself or trigger additional commands in case the
refactoring is happening via something like sourcekit-lsp.

### Modifications:

- Add `addAuxiliaryFiles` and `createAuxiliaryFile` (previously part of
`PackageEdit.applyEdits`) methods to `AddTarget` command to add new
primary and other files to the target directory once the manifest has
been updated.

### Result:

This helps us to remove `ManifestEditRefactoringProvider` (all of the
manifest refactoring actions are going to conform to
`EditRefactoringProvider` protocol directly) and `PackageEdit` types
from swift-syntax and adjust all of the package manifest refactoring
actions to produce `[SourceEdit]` instead.