adds traits to dependency output in show-dependencies --json and --text (#9034)
Adds traits to the output of the show-dependencies command for `--text`
and `--json` output.

### Motivation:

Exposes the default traits, if any, that are enabled by the dependencies
a package.

Resolves: #9033 

### Modifications:

Extended the logic in show-dependencies to add in the `enabledTraits`
for a resolved package.

### Result:

The output of show-dependencies includes output akin to `(traits:
FirstTrait, SecondTrait)` is the resolved package has traits and
includes the traits that it has. If the package doesn't offer traits, no
additional/new content is displayed in `--text`.
The JSON provided from `show-dependencies --json` includes a `traits`
property, empty if there aren't any traits enabled.