Tests: Update helpers to support both test libraries
The `_InternalTestSupport` and `_InternalBuildTestSupport` modules have
XCTest assert helper functions, that, when called from a Swift Testing
test, is essentially a no-op.  During migration to Swift Testing, this
could lead to false positives if the helpers function are not converted
to use Swift Testing APIs.

During the transition, update all helper functions to detect whether the
caller is executing in the XCTest or Swift Testing context, and call the
appropriate API given the library.