Assure default trait is not explicitly added to stored traits dictionary
Since the wrapper EnabledTraitsMap allows us to determine whether
a package is using default traits by returning ["default"] in place
of nil when a package has not yet been added to the stored dict,
omit explicitly adding "default" to the dictionary itself. This will
allow us to check whether we've explicitly modified the trait configuration
for a given package, since it will override the default set of traits.