[Bootstrap] Add an extra argument to set a library path.
During Swift PR testing, we run the SwiftPM tests.  Unfortunately, the
SwiftPM tests always try to link with the system version of the Swift
runtime libraries.  This would be fine, except that Swift Testing has
been built against the *new* runtimes and as a result has ended up
referencing symbols from them that might not be present in the system
runtime libraries.

To fix this, add an extra argument to allow us to add to the
`DYLD_LIBRARY_PATH` setting SwiftPM uses to run its tests.

rdar://160273124