Fix pre-computation of traits when loading dependency manifests (#9057)
This fix addresses some issues that arose with trait-based dependency
resolution on Linux and, found through development, a related issue
found involving the `swift package show-dependencies` command. It fixes
an error that was thrown as a discrepancy in state when executing `swift
package show-dependencies` if traits were guarding a dependency and was
omitted from the package graph.

Additionally, this will assure that the `EnabledTraitsMap` guards
against explicitly adding "default" traits to the stored dictionary,
since this wrapper is a mechanism to determine whether traits have been
explicitly set (and therefore have overridden "default" or have
flattened the list of default traits).