[native] Add toolchain span back deployment rpaths when required (#9164)
SwiftPM's native build system does not have the concept of embedding the
compatibility libraries for distribution, so add rpaths to the span back
deploy library in the toolchain when targeting macOS < 26. Ensure we
always insert /usr/lib/swift ahead of the added rpaths to prefer the OS
content when available. this mirrors the handling of the concurrency
compat lib

Closes https://github.com/swiftlang/swift-package-manager/issues/9163