Use SwiftBuild API to reliably compute built artifacts (#9143)
Depends on https://github.com/swiftlang/swift-build/pull/801

By querying the build system for the artifacts, we ensure we only report
any we will actually produce in the current build. At the same time,
pull in an alternative fix for
https://github.com/swiftlang/swift-package-manager/pull/9121 which
restores the original behavior of the native build system when
considering products in dependency packages.