Only test debug build configuration on Windows
With the recent changes to the automated test, which augmented the Tests
to run against the Native and Swift Build build system, in addition to
the debug and release build configuraiton, and with `swift test` buffer
the output on Windows (#8928), the theory is the Windows self hosted
pipeline is failing due to the number of automated test, which causes
Jenkins to hit the "Build Timeout" algorithim, causing Jenkins to abort
the build as there is no output to the console.

I norer to unblock development, a change was made to run test against
the Debug build configuration on Windows until a proper solution is
availabe.

Related to #9159
Issue: rdar://160936699