[6.2.1][native] Add toolchain span back deployment rpaths when required (#9164) (#9168)
**Explanation**: SwiftPM's native build system does not have the concept
of embedding the compatibility libraries for distribution, so add rpaths
to the span back deploy library in the toolchain when targeting macOS <
26.
- **Scope**: Default package rpaths
- **Issues**:
https://github.com/swiftlang/swift-package-manager/issues/9163
- **Original PRs**:
https://github.com/swiftlang/swift-package-manager/pull/9164
- **Risk**: Low, this change follows an established pattern used by the
concurrency back deployment library
- **Testing**: Updated unit tests, manual testing on older macOS
versions
- **Reviewers**: @jakepetroules 
---
Original:
SwiftPM's native build system does not have the concept of embedding the
compatibility libraries for distribution, so add rpaths to the span back
deploy library in the toolchain when targeting macOS < 26. Ensure we
always insert /usr/lib/swift ahead of the added rpaths to prefer the OS
content when available. this mirrors the handling of the concurrency
compat lib

Closes https://github.com/swiftlang/swift-package-manager/issues/9163