Compile all of the plugins at the start of a build with swiftbuild (#9190)
When building with the swiftbuild build system there can be a
compilation problem with the plugin.swift with one of the plugins.
However, you won't discover that unless you run the command
plugin specifically. This is not the behaviour of the existing native
build system that does a plugin pass at the start of a build.

Implement a similar plugin pass with the swiftbuild system to catch
problems with the plugin code when building the packages.

Fixes #8977