Rename runCommandPlugins methods to a more accurate name (#9223)
This method is very confusing because it suggests that command plugins
will be run during build, which is not the case. Command plugins always
run at the user's request using `swift package plugin <name>` outside of
a build.

Instead, this method should be renamed to reflect that it is running
commands that the plugin has generated when it was invoked from a
collection of results.

Rename the methods in both the SwiftBuildSupport and Build components
and update the documentation.