Show traits package subcommand (#9213)
One of the benefits of traits comes from their declarative nature with a
description that helps to explain their purpose. Users should be able to
discover, and understand them so that they can enable them for the
current package, and also dependencies too so that they can enable them
through the package dependencies.

Add a new show-traits package subcommand that defaults to listing
the traits for the current package, and optionally for other packages
with an option.

    swift package show-traits [--package-id=<packageId>]

Add a format option to set either a text, or JSON output format.

---------

Co-authored-by: Bri Peticca <bripeticca@gmail.com>