extending the docs to explain how to use embedInCode (#9231)
Extends the PackageDescription documentation to explain how to use the
.embedInCode rule for a resource.

### Motivation:

The changelog originally included some detail, but it's missing from the
documentation, so updating the
API reference for PackageDescription to provide information about how to
use .embedInCode.

### Modifications:

Updated the .embedInCode documentation to show code blocks with an
example of usage to illustrate how to use it.

### Result:

Improved documentation for PackageDescription and the embedInCode
resource rule

---------

Co-authored-by: Bri Peticca <bripeticca@gmail.com>