SwiftPM generate-documentation subcommand
Create a subcommand that's capable of generating DocC documentation
for a package, including its targets.

Support natively the ability to document executable targets using
the tool info dump facility available in the Swift Argument Parser
and potentially others that conform to its protocol.

Allow any target regardless of its type to include a DocC catalog
of markdown files to document its target. This includes ones that
don't normally have a mechanism to generate its own API reference.

Provide an option to generate the internal facing documentation
for a Package, including all of its targets, even ones that are
not exported as products.