Fix crash in macros caused by mixing prebuilts and source. (#9261)
There is an issue when a macro that uses prebuilts uses a library that
also depends on the swift-syntax library. This is commonly used to share
utility code between macros.

However, we currently can't tell whether the library will only be built
for the host so it can not take advantage of the prebuilts. When a macro
which is using prebuilts uses that library, we are seeing crashes in the
macro caused when the library and the swift-syntax it uses is built for
debug which is then linked with the macro which uses the swift-syntax
prebuilts which are built for release.

For now, this change turns off the prebuilts for macros that use such a
library to avoid the crash. We'll investigate making the use of
prebuilts conditional on the build settings in a future release.

To be able to implement this change, we need to be able to detect the
downstream swift-syntax dependency. We move the insertion of the
settings for the prebuilts from Module creation time until after the
ResolvedModuleBuilder is created so that we can make the changes there
before the final ResolvedModules are created.