Bootstrap: Use computed module triple
The target triple coming from `get_build_target` is either the
unversioned triple or the full target triple. Neither of which is the
module triple, which comes from the `moduleTriple` field of the output
coming from the compiler. Use the `moduleTriple` field to compute the
module triple for use in the filename.