Fix for required deps calculation with traits
The enabled traits property was essentially being erased to defaults
rather than propagating what was calculated earlier due to how we
were returning the DependencyResolutionNode during the pub grub
stage.

This fix should now assure that we are passing the appropriate
enabled traits to these nodes, and are being considered when making
calls to PackageContainer's dependency-related methods.