[6.2][cherrypick]Add bootstrap option to install without building. (#9105)
- **Explanation**:

This allows the build action to be decoupled from the install action, so
packaging systems that expect to be able to do discrete installation
steps post-build can do so cheaply, without having to effectively start
the build over from scratch.

This is only intended solely for that use-case, so this is not wired up
to other part of the swiftpm build process and defaults to being
disabled.

(See also #9106)

- **Scope**:

Makes a change to the bootstrap wrapper, so no behavioral changes by
default.

- **Issues**:

Relevant to OpenBSD port in swiftlang/swift#78437

- **Original PRs**:

#9050 

- **Risk**:

Minimal, since there are no behavioral changes by default.

- **Testing**:

Passed CI.

- **Reviewers**:

@jakepetroules