Detect system GNU ld script pretending to be system libraries
Basic C runtime libraries appear to sometimes be replaced by a GNU ld
script that pulls it some other libraries as a backwards compatibility
mechanism on certain Linux distros. The usage of this workaround doesn't
seem stable across distro versions so this commits adds support for the
subset that of GNU ld scripts that are used to package such libraries.