Add an experimental interface for custom test libraries.
This PR adds command-line experimental support for custom test libraries that don't sit on top of XCTest or Swift Testing (irony inbound). Developers can specify additional libraries to invoke using the `--experimental-testing-library` argument.

In a separate PR, I will teach Swift Testing how to read the `--experimental-testing-library` argument (forwarded as `--testing-library` once per, er, library.) A developer can write a small shim function that gets added to the `__swift5_tests` section as a special test content record and which, when called, invokes its corresponding testing library and translates output to the "language" Swift Testing speaks.

The above description is light on details because we haven't designed out the mechanisms in question. In particular, we don't have a stable configuration/input mechanism, so for the moment I'm just forwarding command-line arguments. We also need to teach some part of the stack how to collate results from multiple libraries to produce a single unified test report in various formats. This PR is just one step.