Modify emission of command line display strings
If there is no command line display strings available for a failed
task, we should demote this message to info-level to avoid cluttering
the output stream with messages that may not be incredibly helpful.

To consider here: if this is the only error, we should be able to expose
this as an error and perhaps omit "<no command line>".